Implement command for confirming command sequences.
First attempt at implementing a confirmation command intended for use in
command sequences.

When called (optionally with a free-form string message), prompts the
user for confirmation. If the user declines, a SequenceCanceled
exception is raised, which trickles up the error handling into
apply_commandline where it is explicitly caught outside the common error
handler. This aborts the execution of subsequent commands in the
commandline, but does not show up as an error in the log.

Closes #1368