db/utils: perform signature validation on original message bytes
Parsing and re-serialising a Message sometimes alters the original content
(by removing extra lines between nested message multiparts for example).
This interfers with the signature validation which requires the byte content
to match perfectly.

To avoid this issue, we perform the signature validation on the original
message bytes directly.

GitHub: resolves #1459