Don't lose the focus when refreshing a thread
This commit allows keeping the currently focused message selected
across thread refreshes.

Since this operation can take a long time in large threads, a
`search_threads_rebuild_limit` option is introduced. When defined,
it sets the maximum amount of messages that will be iterated upon
when trying to find the previously focused message.