PipeCommand: pass shell arg to ExternalCommand.
Prior to this branch, `shell=True` was passed to Popen regardless of
whether `--shell` was passed. I'm not sure why this was the case but it
wasn't a problem. Now that we're passing to ExternalCommand, this
doesn't work because ExternalCommand does it's own manipulation of cmd
when shell=False which is incompatible with a list-style cmd.