taglist: match by regular expression
Currently, there is no way to specify the taglist filter function from
UI, and the tags argument just allows to display the list of tags that
you specify.

Introduce a match argument that allows to list those tags matching the
specified regular expression.

Especially useful if you bind a key to `prompt 'taglist '`.