taglist: refine search from buffer
When a taglist buffer lists tags from a search or thead buffer (without
`--global`) then selecting a tag lists all messages in the db with that
tag. This feels counter-intuitive.

Instead, remember the original query and refine the query by filtering
on the tag in addition to the original query when the tag is `select`ed.

In addition, introduce a new command `globalselect` which does a global
search with the selected tag. The default key binding is `meta enter`,
complementing the default binding `enter` for `select`.