taglist buffer: make statusbar informative
Until recently, taglist buffers were global. Now they can differ by
realm (global, search, thread) and pattern matching on the tags. Provide
this information for the statusbar and update the defaults to display
it.

In addition, provide a count for the messages matching the specific
query, i.e. globally or the query from the search or thread buffer,
restricted to matches on the pattern matched tags.

Note that even in the global case without pattern match the matched
count is typically lower than the total count since the latter includes
messages with excluded tags as well as those without any tags, whereas
the former list any not excluded message with any tag.