taglist: show message counts
So far, the taglist shows a list of tags (no surprise here) only,
recently with a total count of matched message.

Show a count for each tag analogous to the namelist widget:
These are the counts for original query (global, search, or thread) AND
the tag, and also for that AND tag:unread.