taglistbuffer: deal with 0 tags
If a taglist buffer is asked to display 0 tags then it throws a
ZeroDivisionError currently. Since people usually have at least 1 tag
nobody noticed so far. But in the future the UI will allow to restrict
the set of tags in a taglist buffer, so prepare for the case of 0 tags.