taglist: do not quote tag unnecessarily
Currently, selecting a tag searches for that tag surrounded by quotes.
This results in a somewhat strange display in the resulting search
buffer status line.

Surround the tag by quotes only if it contains a space. We do the same
when preparing the input line for `retag`, for example (where it is
ncessary), and it gives a more natural display.