tag commands: provide --fullthread option
Since the search buffer is a view to `notmuch search`, the tag commands
act on matched messages in the same ways as `notmuch tag` does.

Implement an option `--fullthread` that makes the tag commands act on
all messages in a thread instead. It's often the more useful variant in
the UI, and this was the behaviour when alot treated the search buffer
as a buffer of threads (as opposed to a buffer of matched messages).