toggletags: restore correct behaviour in a search buffer
`toggletags` is supposed to toggle tags based on the presence of a tag
in a thread or search result.

0cd98062 ("db/manager: Handle tag toggling messages", 2020-05-18) moved
the handling of `toggletags` (i.e. the determination of tags to remove
and add) from the search buffer command to the db manager. Alas, it also
turned the handling to be per message: If you toggle the `unread` tag on
a thread with unread messages then the unread ones get untagged but the
read messages get tagged `unread`.

This might have gone unnoticed so far because - as long as the search
buffer does not get requeried by thread IDs - a search buffer for
`tag:unread` would not match other messages and would not act on them at
all. But this is different when you toggle a tag that you have not
searched for, of course.

(Re-)Teach the db manager to toggle tags based on the presence of a tag
in the results of a query.