toggletags: refactor and fix behaviour in a thread buffer
0cd98062 ("db/manager: Handle tag toggling messages", 2020-05-18) moved
the handling of `toggletags` (i.e. the determination of tags to
removeand add) from the search buffer command to the db manager. It did
not move the corresponding code from the thread buffer command, though.

Refactor the thread buffer command so that it calls the db manager
methods as the search buffer command does. Besides reducing reduncancy,
this also fixes an old issue with `toggletags --all` in a thread buffer:
If this is supposed to do something useful and unsursprising then it
should act the same way as `toggletags` in a search buffer for
`thread:tid`, i.e. toggle a tag based on the presence in the thread (not
individually). The refactoring achieves this as an intended side-effect.