toggletags: provide fullthread option alias
Depending on your point of view, `--all` means something different or
the same for the tag commands in search and thread buffers.

Provide an alias `--fullthread` for the `--all` option of tag commands
in a thread buffer so that one can use the same key binding for acting
on one full thread in either buffer.