db: message: Implement get_subject
Provide a method to get the subject of a given message.

The subject is cached by not-much in the database, and is efficient to
access directly through the notmuch2.msg object.

Store the subject in our Message representation so it can be used
efficiently when parsing and displaying mail threads.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>

---
Fix no subject mails