commands: thread: Reference existing message when forwarding
When forwarding a mail, the outgoing mail, and any follow up
conversation becomes a disjointed thread.  This removes the original
context, and makes it harder to find and track while the conversation
splits in two directions.

Add appropriate reference headers to the envelope when constructing a
forwarded mail to ensure that the conversations continue to be related
and therefore presented in the same thread with notmuch and alot.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>