Invert mailcap logic: x-alot-attachmenttotext
Rather than use x-alot-openattachment, and rely on the default ('view')
mailcap entry to dump the attachment to text, assume that the default
entry is set up to handle attachments *as attachments*, i.e. open them
in whatever application is supposed to handle them.

Instead, define x-alot-attachmenttotext that provides a way to render
the attachment as text. If defined, opening the attachment through the
mime tree will both show it as text *and* open it. If not defined,
opening the attachment through the mime tree will only open it.

When opening the attachment through the attachment widget, it's still
only opened even if an x-alot-attachmenttotext handler is defined.