Make sure that we have a valid path for DBManager
Otherwise, we run into problems when we need to use the path attribute
and it is None (for example, in DBManager.add_message()).

Getting the path directly from the notmuch config file is not enough,
because the user might be using the default location with no explicit
entry in the file.

Since there is a bit of logic notmuch does to get the database path (see
section "DATABASE LOCATION" in man page for notmuch-config(1)), it is
better to ask it directly via "notmuch config" instead of reimplementing
that same logic.