alot/settings: Use notmuch config to read configuration
The main benefit from this is that we get the defaults directly from
notmuch and do not risk using inconsistent configuration values.

A good example of this is getting the database path. Getting the path
directly from the notmuch config file is not enough, because the user
might be using the default location with no explicit entry in the file.
Since notmuch does is a bit of logic notmuch to get the database path
(see section "DATABASE LOCATION" in man page for notmuch-config(1)), it
is better to get it directly from "notmuch config".

Since get_notmuch_setting() now always returns a string, also refactor
the call to settings.get_notmuch_setting('maildir', 'synchronize_flags')
for this new reality.