commands/thread: Add option --strip_ansi to pipeto
If someone has a mailcap entry that colors the mail content, it is
possible that she would want to get the content without the CSI escapes
when piping it an external command.

Ideally, there should be a way to have the content untouched by the
mailcap filter so we could use it when it makes sense. However, let's
use ansi.remove_csi() for the time being.