bump gpg dependency
There is a problamatically old version of the official gpgme bindings on
pypi (https://pypi.org/project/gpg/)
which gets pulled by defaul if one installs alot via pip.

This commit introduces a dependency on gpg at strictly higher version to
avoid this outdated dependency.

It is likely that this will cause people some pain because there is no
alternative installable via pip as far as I know.
However, several distributions already ship with the newer official
bindings (python3-gpg) on debian and derivatives, which means it will
not be an issue there.