editnew: Use raw body text instead of rendered version
Rendering mail content via mailcap makes sense when viewing email, but
not when reusing the email for composing a new message, as it is
desirable to rather edit the "original" content and, if mailcap
rendering produces ANSI CSI escapes, the user would be left with a bunch
of garbarge characters.

Make sure that the editnew command do not render the body when creating
the envelope.