Make hooks optional
This changes the default setting of the 'hooksfile' setting
and associated behaviour so that, unless explicitly set,
this config setting defaults to `None`, which then means loading the
hooksfile is not even attempted.

Prior to this commit, the default value was
`$XDG_CONFIG_HOME/alot/hooks.py`
and no alot would try to read that file upon start (and report via
logging.exception if it failed, which in the ui is uncritical).

This was problematic not only in that it hardcodes the unix separator
'/' but also because it caused unecessary disk access / errors.
More importantly, it caused several test cases to fail (see
[here](https://bugs.gentoo.org/874372#c2).

The patch also includes updated docs.