commands/search/tag: Narrow search down to needed messages
For some reason, tagging long threads is taking a lot of time in my
system. It might be related to the python library for notmuch, since
doing the same directly from notmuch CLI is usually fast.

In some use cases (e.g. archiving), most of the messages in the thread
might already contain the tags being added or removed. We can optimize
for this case by narrowing down the search to only the messages that
need the update.

So, while the slow tagging operation should be properly investigated at
the library level, we can proactively optimize the testquery,
which is beneficial anyway.