db/utils: correctly handle non-UTF8 8bit encodings
The 8bit encoded mail parts should be encoded consistently with the
charset they specify in their Content-Type header. Otherwise subsequent
decoding of non-ASCII characters is likely to produce the wrong string.

Currently our usage of [Message.get_payload] returns a string decoded
from the charset in the message header. From discussion in pazz#1522,
pazz#1526 and 777823f it seems that it its behavior (or our usage of it)
may have changed over the years. In the 8bit case this code essentially
tries to undo the get_payload decoding. Arguably it would be better if
the email library provided a proper bytestring or its internal ASCII
with surrogateescape representation, but this is what we have available.

[Message.get_payload]: https://github.com/python/cpython/blob/96b7a2eba423b42320f15fd4974740e3e930bb8b/Lib/email/message.py#L297